Reorganization of code in feat/, moving things between files for better division by topic.
