+++ src/feat/feature-fbank.h
@@ -24,13 +24,15 @@
 #include <string>
 
 #include "feat/feature-functions.h"
+#include "feat/feature-window.h"
+#include "feat/mel-computations.h"
 
 namespace kaldi {
 /// @addtogroup  feat FeatureExtraction
 /// @{
 
 
-/// FbankOptions contains basic options for computing FBANK features
+/// FbankOptions contains basic options for computing filterbank features.
 /// It only includes things that can be done in a "stateless" way, i.e.
 /// it does not include energy max-normalization.
 /// It does not include delta computation.
