+++ src/feat/mel-computations.h
@@ -1,6 +1,7 @@
 // feat/mel-computations.h
 
 // Copyright 2009-2011  Phonexia s.r.o.;  Microsoft Corporation
+//                2016  Johns Hopkins University (author: Daniel Povey)
 
 // See ../../COPYING for clarification regarding multiple authors
 //
@@ -36,9 +37,43 @@ namespace kaldi {
 /// @addtogroup  feat FeatureExtraction
 /// @{
 
-struct FrameExtractionOptions;  // defined in feature-function.h
+struct FrameExtractionOptions;  // defined in feature-window.h
 
-struct MelBanksOptions;  // defined in feature-function.h
+
+struct MelBanksOptions {
+  int32 num_bins;  // e.g. 25; number of triangular bins
+  BaseFloat low_freq;  // e.g. 20; lower frequency cutoff
+  BaseFloat high_freq;  // an upper frequency cutoff; 0 -> no cutoff, negative
+  // ->added to the Nyquist frequency to get the cutoff.
+  BaseFloat vtln_low;  // vtln lower cutoff of warping function.
+  BaseFloat vtln_high;  // vtln upper cutoff of warping function: if negative, added
+                        // to the Nyquist frequency to get the cutoff.
+  bool debug_mel;
+  // htk_mode is a "hidden" config, it does not show up on command line.
+  // Enables more exact compatibibility with HTK, for testing purposes.  Affects
+  // mel-energy flooring and reproduces a bug in HTK.
+  bool htk_mode;
+  explicit MelBanksOptions(int num_bins = 25)
+      : num_bins(num_bins), low_freq(20), high_freq(0), vtln_low(100),
+        vtln_high(-500), debug_mel(false), htk_mode(false) {}
+
+  void Register(OptionsItf *opts) {
+    opts->Register("num-mel-bins", &num_bins,
+                   "Number of triangular mel-frequency bins");
+    opts->Register("low-freq", &low_freq,
+                   "Low cutoff frequency for mel bins");
+    opts->Register("high-freq", &high_freq,
+                   "High cutoff frequency for mel bins (if < 0, offset from Nyquist)");
+    opts->Register("vtln-low", &vtln_low,
+                   "Low inflection point in piecewise linear VTLN warping function");
+    opts->Register("vtln-high", &vtln_high,
+                   "High inflection point in piecewise linear VTLN warping function"
+                   " (if negative, offset from high-mel-freq");
+    opts->Register("debug-mel", &debug_mel,
+                   "Print out debugging information for mel bin computation");
+  }
+};
+
 
 class MelBanks {
  public:
@@ -109,7 +144,16 @@ void ComputeLifterCoeffs(BaseFloat Q, VectorBase<BaseFloat> *coeffs);
 //       F(z) = 1 / (1 - A(z)), 1 is not stored in the demoninator
 BaseFloat Durbin(int n, const BaseFloat *pAC, BaseFloat *pLP, BaseFloat *pTmp);
 
+// Compute LP coefficients from autocorrelation coefficients.
+BaseFloat ComputeLpc(const VectorBase<BaseFloat> &autocorr_in,
+                     Vector<BaseFloat> *lpc_out);
+
 void Lpc2Cepstrum(int n, const BaseFloat *pLPC, BaseFloat *pCepst);
+
+
+
+void GetEqualLoudnessVector(const MelBanks &mel_banks,
+                            Vector<BaseFloat> *ans);
 
 /// @} End of "addtogroup feat"
 }  // namespace kaldi
